smart_asset/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
axios==0.4.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.2
colorama==0.4.6
distlib==0.4.0
Django==5.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
filelock==3.25.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
# lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
# mysql-connector-python==9.5.0
# ollama==0.6.1
packaging==26.0
pillow==12.1.1
platformdirs==4.9.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
python-discovery==1.1.0
requests==2.33.1
rich==14.3.3
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
virtualenv==21.1.0
whitenoise==6.12.0
dj-database-url